requirements.txt
appnope==0.1.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
bs4==0.0.1
certifi==2016.2.28
cycler==0.10.0
decorator==4.0.9
dnspython==1.12.0
docutils==0.12
email-validator==1.0.1
Flask==0.10.1
Flask-SQLAlchemy==2.1
functools32==3.2.3.post2
gnureadline==6.3.3
idna==2.1
ipython==4.2.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
lockfile==0.12.2
luigi==2.1.1
MarkupSafe==0.23
matplotlib==1.5.1
mysql==0.0.1
mysql-connector-python-rf==2.0.4
MySQL-python==1.2.5
numpy==1.11.0
pandas==0.18.1
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
ptyprocess==0.5.1
pyparsing==2.1.5
python-daemon==2.1.1
python-dateutil==2.5.3
pytz==2016.4
requests==2.10.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.0.12
subprocess32==3.2.7
tornado==4.3
traitlets==4.2.1
Werkzeug==0.11.9